+++ lib/components/Icon/index.tsx
@@ -1,9 +1,10 @@
-import { CSSProperties } from 'react'
+import { CSSProperties, useMemo, useState } from 'react'
 import classnames from 'classnames'
 
 import useAtomic from '@/hooks/useAtomic'
 
 import useStyles from './styles'
+import iconSizes from '@/tokens/iconSizes'
 
 export interface IconProps {
   iconName: string
@@ -19,10 +20,26 @@ export interface IconProps {
   testId?: string
 }
 
-const Icon = (props: IconProps) => {
-  const classes = useStyles(props)
+const Icon = ({
+  size,
+  color,
+  hoverColor,
+  transition,
+  iconName,
+  className,
+  style,
+  id,
+  onClick,
+  testId
+}: IconProps) => {
+  const [hover, setHover] = useState(false)
+  const classes = useStyles()
   const atomic = useAtomic()
-  const { iconName, className, style, id, onClick, testId } = props
+  const iconColor = useMemo(() => {
+    if (hover && hoverColor) return hoverColor
+    if (color) return color
+    return 'currentcolor'
+  }, [color, hoverColor, hover])
   return (
     <svg
       className={classnames(
@@ -33,12 +50,22 @@ const Icon = (props: IconProps) => {
         },
         className
       )}
+      width={size ? size : iconSizes.base}
+      height={size ? size : iconSizes.base}
+      fill={iconColor}
       id={id}
-      style={style}
+      style={{
+        ...style,
+        transition: transition ? transition : '0.3s all'
+      }}
       onClick={onClick}
+      onMouseEnter={() => setHover(true)}
+      onMouseLeave={() => setHover(false)}
       data-testid={testId}
     >
-      <use xlinkHref={`#${atomic.iconsPrefix}__${iconName}`} />
+      <use
+        xlinkHref={`${atomic.iconsPath}#${atomic.iconsPrefix}__${iconName}`}
+      />
     </svg>
   )
 }