+++ lib/components/NavTab/index.tsx
@@ -2,8 +2,8 @@ import React, {
   useState,
   useEffect,
   useCallback,
-  useMemo,
-  ReactNode
+  ReactNode,
+  SyntheticEvent
 } from 'react'
 import classnames from 'classnames'
 
@@ -12,41 +12,82 @@ import Flexbox from '@/components/Flexbox'
 import NavItem from '@/components/NavItem'
 import NavIcon from '@/components/NavIcon'
 import NavTop, { TopProps } from '@/components/NavTop'
-import Button from '@/components/Button'
+import Button, { ButtonTheme } from '@/components/Button'
 import Icon from '@/components/Icon'
 import colors from '@/tokens/colors'
 import spacing from '@/tokens/spacing'
-import grid from '@/tokens/grid'
 import iconSizes from '@/tokens/iconSizes'
-import useWindowSize from '@/hooks/useWindowSize'
 
 import useStyles from './styles'
 
-interface NavElement {
+export interface LinkElement {
   key: string | number
-  type: 'button' | 'link' | 'dropdownLink' | 'icon' | 'custom' | 'logo'
-  text?: string
-  label?: string
-  onClick?: () => void
+  type: 'link'
+  text: string
+  onClick?: (e: SyntheticEvent) => void
   selected?: boolean
   link?: string
-  theme?:
-    | 'primary'
-    | 'secondary'
-    | 'tertiary'
-    | 'tertiaryWhite'
-    | 'ghostPink'
-    | 'ghostGrey'
-    | 'ghostWhite'
-  iconName?: string
-  showBar?: boolean
-  custom?: ReactNode
   testId?: string
 }
 
+export interface ButtonElement {
+  key: string | number
+  type: 'button'
+  text?: string
+  onClick?: (e: SyntheticEvent) => void
+  theme?: ButtonTheme
+  iconName?: string
+  testId?: string
+}
+
+export interface DropdownElement {
+  key: string | number
+  type: 'dropdownLink'
+  text: string
+  onClick?: (e: SyntheticEvent) => void
+  selected?: boolean
+  link?: string
+  testId?: string
+}
+
+export interface IconElement {
+  key: string | number
+  type: 'icon'
+  label?: string
+  onClick?: (e: SyntheticEvent) => void
+  selected?: boolean
+  iconName: string
+  testId?: string
+  showBar?: boolean
+  direction?: 'col' | 'row'
+  width?: number
+}
+
+export interface CustomElement {
+  key: string | number
+  type: 'custom'
+  custom: ReactNode
+  testId?: string
+}
+
+export interface LogoElement {
+  key: string | number
+  type: 'logo'
+  logo: ReactNode
+  testId?: string
+}
+
+export type NavElement =
+  | LinkElement
+  | ButtonElement
+  | DropdownElement
+  | IconElement
+  | CustomElement
+  | LogoElement
+
 export interface NavPosition extends Array<NavElement> {}
 
-interface NavTabProps {
+export interface NavTabProps {
   left?: NavPosition
   right?: NavPosition
   center?: NavPosition
@@ -58,6 +99,8 @@ interface NavTabProps {
   bottom?: boolean
   zIndex?: number
   className?: string
+  /** The recommendation is to set the breakpoint at `grid.xl` */
+  isFluid?: boolean
 }
 
 const NavTab = ({
@@ -71,12 +114,12 @@ const NavTab = ({
   bottom,
   className,
   zIndex,
-  hideOnScroll
+  hideOnScroll,
+  isFluid
 }: NavTabProps) => {
   const classes = useStyles()
   const [show, setShow] = useState(true)
   const [currentScroll, setCurrentScroll] = useState(0)
-  const { width } = useWindowSize()
 
   const determineVisibility = useCallback(() => {
     setShow(
@@ -97,7 +140,7 @@ const NavTab = ({
   }, [fixed, hideOnScroll, determineVisibility])
 
   const renderLink = useCallback(
-    item => {
+    (item: LinkElement) => {
       return (
         <NavItem white={blue} {...item} className={classes.navItem}>
           {item.text}
@@ -108,7 +151,7 @@ const NavTab = ({
   )
 
   const renderDropdownLink = useCallback(
-    item => {
+    (item: DropdownElement) => {
       return (
         <NavItem
           white={blue}
@@ -132,7 +175,7 @@ const NavTab = ({
   )
 
   const renderButton = useCallback(
-    item => {
+    (item: ButtonElement) => {
       return (
         <Button className={classes.button} {...item}>
           {item.text}
@@ -143,7 +186,7 @@ const NavTab = ({
   )
 
   const renderIcon = useCallback(
-    item => {
+    (item: IconElement) => {
       return (
         <div className={classes.iconWrap} key={item.key}>
           <NavIcon className={classes.icon} white={blue} {...item} />
@@ -154,9 +197,9 @@ const NavTab = ({
   )
 
   const renderLogo = useCallback(
-    item => {
+    (item: LogoElement) => {
       return (
-        <div className={classes.logo} key={item.key}>
+        <div data-testid={item.testId} className={classes.logo} key={item.key}>
           {item.logo}
         </div>
       )
@@ -164,18 +207,30 @@ const NavTab = ({
     [classes]
   )
 
-  const renderCustom = useCallback(item => {
-    return <div key={item.key}>{item.custom}</div>
+  const renderCustom = useCallback((item: CustomElement) => {
+    return (
+      <div data-testid={item.testId} key={item.key}>
+        {item.custom}
+      </div>
+    )
   }, [])
 
   const renderItem = useCallback(
-    item => {
-      if (item.type == 'link') return renderLink(item)
-      else if (item.type == 'dropdownLink') return renderDropdownLink(item)
-      else if (item.type == 'button') return renderButton(item)
-      else if (item.type == 'icon') return renderIcon(item)
-      else if (item.type == 'logo') return renderLogo(item)
-      else if (item.type == 'custom') return renderCustom(item)
+    (item: NavElement) => {
+      switch (item.type) {
+        case 'link':
+          return renderLink(item)
+        case 'dropdownLink':
+          return renderDropdownLink(item)
+        case 'button':
+          return renderButton(item)
+        case 'icon':
+          return renderIcon(item)
+        case 'logo':
+          return renderLogo(item)
+        case 'custom':
+          return renderCustom(item)
+      }
     },
     [
       renderLink,
@@ -187,7 +242,6 @@ const NavTab = ({
     ]
   )
 
-  const isFluid = useMemo(() => width < grid.xl, [width])
   return (
     <div
       className={classnames(