+++ src/components/Header/index.tsx
@@ -1,12 +1,13 @@
 import { useState } from 'react'
 import Link from 'next/link'
 
-import NavTab from '@/components/NavTab'
+import NavTab, { NavTabProps } from '@/components/NavTab'
 import NavAside from '@/components/NavAside'
 import Portal from '@/components/Portal'
 import Menu from '@/src/components/Menu'
 import Search from '@/src/components/Search'
 import { Doc } from '@/src/components/Layout'
+import { Media } from '@/components/Media'
 
 import useStyles from './styles'
 
@@ -18,44 +19,49 @@ export default function Header({ docs }: HeaderProps) {
   const classes = useStyles()
   const [showAside, setShowAside] = useState(false)
 
+  const navTabContent: NavTabProps = {
+    left: [
+      {
+        key: 'icon',
+        type: 'icon',
+        iconName: 'menu',
+        onClick: () => setShowAside(true)
+      },
+      {
+        key: 'custom',
+        type: 'custom',
+        custom: (
+          <Link href="/">
+            <a className={classes.title}>Atomic</a>
+          </Link>
+        )
+      }
+    ],
+    right: [
+      {
+        key: 'link',
+        type: 'link',
+        link: 'https://github.com/occmundial/atomic',
+        text: 'View on Github'
+      }
+    ],
+    center: [
+      {
+        key: 'custom2',
+        type: 'custom',
+        custom: <Search docs={docs} />
+      }
+    ]
+  }
+
   return (
     <>
-      <NavTab
-        fixed
-        hideOnScroll
-        left={[
-          {
-            key: 0,
-            type: 'icon',
-            iconName: 'menu',
-            onClick: () => setShowAside(true)
-          },
-          {
-            key: 1,
-            type: 'custom',
-            custom: (
-              <Link href="/">
-                <a className={classes.title}>Atomic</a>
-              </Link>
-            )
-          }
-        ]}
-        right={[
-          {
-            key: 0,
-            type: 'link',
-            link: 'https://github.com/occmundial/atomic',
-            text: 'View on Github'
-          }
-        ]}
-        center={[
-          {
-            key: 0,
-            type: 'custom',
-            custom: <Search docs={docs} />
-          }
-        ]}
-      />
+      <Media lessThan="xs">
+        <NavTab fixed hideOnScroll {...navTabContent} isFluid />
+      </Media>
+      <Media greaterThanOrEqual="xs">
+        <NavTab fixed hideOnScroll {...navTabContent} isFluid={false} />
+      </Media>
       <Portal show={showAside}>
         <NavAside
           onClose={() => setShowAside(false)}